+++ pyoffers/models/raw_log.py
@@ -3,6 +3,8 @@ from csv import DictReader
 from io import BytesIO, StringIO
 from zipfile import ZipFile
 
+from pip.utils import cached_property
+
 from .core import Model, ModelManager
 
 
@@ -35,11 +37,11 @@ class LogFile(LogItem):
     def __str__(self):
         return '%s: %s (%s)' % (self.__class__.__name__, self.displayName, self.filename)
 
-    @property
+    @cached_property
     def download_link(self):
         return self._manager.get_download_link(log_filename=self.filename)
 
-    @property
+    @cached_property
     def content(self):
         """
         Returns raw CSV content of the log file.
@@ -50,7 +52,7 @@ class LogFile(LogItem):
         filename = archive.filelist[0]  # Always 1 file in the archive
         return archive.read(filename)
 
-    @property
+    @cached_property
     def records(self):
         data = StringIO(self.content.decode())
         return [LogRecord(manager=self._manager, **kwargs) for kwargs in DictReader(data)]